refactor: rename infrastructure to data_providers
