+++ api/src/features/todo/use_cases/add_todo.py
@@ -2,8 +2,10 @@ import uuid
 
 from pydantic import BaseModel, Field
 
+from data_providers.repository_interfaces.TodoRepositoryInterface import (
+    TodoRepositoryInterface,
+)
 from entities.TodoItem import TodoItem
-from features.todo.interfaces.TodoRepositoryInterface import TodoRepositoryInterface
 
 
 class AddTodoRequest(BaseModel):
