+++ api/src/features/todo/use_cases/delete_todo_by_id.py
@@ -1,6 +1,8 @@
 from pydantic import BaseModel, Field
 
-from features.todo.interfaces.TodoRepositoryInterface import TodoRepositoryInterface
+from data_providers.repository_interfaces.TodoRepositoryInterface import (
+    TodoRepositoryInterface,
+)
 
 
 class DeleteTodoByIdResponse(BaseModel):
