+++ api/src/features/todo/use_cases/get_todo_all.py
@@ -2,8 +2,10 @@ from typing import List
 
 from pydantic import BaseModel, Field
 
+from data_providers.repository_interfaces.TodoRepositoryInterface import (
+    TodoRepositoryInterface,
+)
 from entities.TodoItem import TodoItem
-from features.todo.interfaces.TodoRepositoryInterface import TodoRepositoryInterface
 
 
 class GetTodoAllResponse(BaseModel):
