+++ api/src/tests/unit/infrastructure/clients/mongodb/test_MongoDatabaseClient.py
@@ -1,7 +1,7 @@
 import pytest
 
 from common.exceptions import NotFoundException, ValidationException
-from infrastructure.clients.mongodb.MongoDatabaseClient import MongoDatabaseClient
+from data_providers.clients.mongodb.MongoDatabaseClient import MongoDatabaseClient
 
 
 class TestMongoDatabaseClient:
