+++ documentation/docs/contributing/development-guide/coding/extending-the-api/adding-data-providers/index.md
@@ -3,8 +3,9 @@
 Data providers are part of the infrastructure layer, which is responsible for external infrastructure communications like database storage, file system, and external systems. The infrastructure layer is the layer that contains all the concrete implementations of the application. It implements interfaces defined in use cases, to provide access to external systems. 
 
 ```
-├── data-providers/
+├── data_providers/
 │   ├── clients/ 
+│   ├── repository_interfaces/ 
 │   └── repositories/
 └── ...
 ```
