+++ documentation/docs/contributing/development-guide/coding/extending-the-api/adding-features/index.md
@@ -4,7 +4,6 @@ A feature has this structure.
 
 ```
 ├── todo/
-│   ├── interfaces/ - Repository interfaces (optional)
 │   ├── use_cases/ - Application logic
 │   ├── exceptions.py - Exceptions classes (optional)
 │   └── controller.py - The entrypoint
